requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.0
iorodeo-potentiostat==0.0.9
kiwisolver==1.4.8
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
progressbar33==2.4
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
-e git+https://github.com/benjaminrodriguezonline/Radiostat_Testing.git@547f330f1bbd45bca884b1f42e02cce04c16f2a3#egg=runRadiostat
six==1.17.0
tzdata==2025.2
